ril: Fix data after reboot
In previous versions, setPreferredNetworkType was called directly after
startup while the modem was powered down so the correct network type
was available while the modem is initializing. This changed in M
and the preferred network type is set much later which causes the
modem to not allow datacalls until it is rebooted again.
Rebooting the modem the first time setPreferredNetworkType is
encountered fixes this.

Change-Id: I8068909f42b8b4de484968f7088c51e9f99affef